Completely finished setup function
Finished the rest of the required information needed to setup the player. Made minor changes to print statements so the correct info is printed (such as gender). Made the text more spread out so it is easier to read. Added the ability to remake the character at the end of the setup if the player made a mistake in the process.